feat: use a theme_id (position of the theme in the lua table) to determine what the current theme is (#5)
When using themes that have different variations/flavors that are set using global variables, the cursor is not placed on the correct line in the menu buffer when opened. Similarly, the '>' is not always added to the beggining of the appropriate theme. This introduces a currentThemeId, which replaces the need for both currentThemeName and currentThemeIndex in controller.lua, while also leading to more accurate results.

Co-authored-by: Juan Torres <imzaldih@gmail.com>